Moved admin's action view tests to a separate file.
